myprojectbe/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
xlrd==2.0.1
lxml==5.2.1

python-dotenv==1.0.1
supabase==2.6.0

passlib[bcrypt]==1.7.4
pyjwt==2.8.0